# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ashva Experts API (Consolidated)
  version: 0.2.0
  description: |
    Consolidated API specification for manual testing during development.

    Notes:
    - Base URL for local testing via Nginx: `http://localhost:8080`
    - All `/api/v1/*` routes are served through the gateway.
    - Auth endpoints under `/api/v1/auth/*` are public.
    - Non-auth endpoints require `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:8080
    description: Local via Nginx (recommended)
  - url: http://localhost:8000
    description: Local gateway direct
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
...
    SubscriberMeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        company_name: { type: string, nullable: true }
        gst_number: { type: string, nullable: true }
        pan_number: { type: string, nullable: true }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, user_id, first_name, last_name, email, is_active, created_at, updated_at]
    SubscriberUpdateRequest:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        phone: { type: string }
        company_name: { type: string }
        gst_number: { type: string }
        pan_number: { type: string }
...
    TicketCreateRequest:
      type: object
      properties:
        ticket_type: { type: string, enum: [installation, service, repair] }
        priority: { type: string, enum: [low, medium, high, urgent], default: medium }
        title: { type: string }
        description: { type: string }
        location_address: { type: string, nullable: true }
      required: [ticket_type, title, description]
    TicketResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        ticket_number: { type: string }
        subscriber_id: { type: string, format: uuid }
        subscription_id: { type: string, format: uuid, nullable: true }
        ticket_type: { type: string, enum: [installation, service, repair] }
        priority: { type: string, enum: [low, medium, high, urgent] }
        status: { type: string, enum: [created, assigned, in_progress, completed, closed, cancelled] }
        title: { type: string }
        description: { type: string }
        location_address: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required:
        - id
        - ticket_number
        - subscriber_id
        - ticket_type
        - priority
        - status
        - title
        - description
        - created_at
        - updated_at
    TicketListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TicketResponse" }
        total: { type: integer }
      required: [items, total]
...
paths:
...
  /api/v1/subscribers/me:
    get:
      tags: [subscribers]
      summary: Get current subscriber profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriberMeResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [subscribers]
      summary: Update current subscriber profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriberUpdateRequest" }
            example:
              first_name: John
              last_name: Doe
              company_name: Acme Corp
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriberMeResponse" }
...
  /api/v1/tickets:
    post:
      tags: [tickets]
      summary: Create a ticket (subscriber)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketCreateRequest" }
            example:
              ticket_type: service
              priority: medium
              title: "AC not cooling properly"
              description: "AC unit is running but not cooling the room"
              location_address: "123 Main St, Mumbai"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketResponse" }
  /api/v1/tickets/me:
    get:
      tags: [tickets]
      summary: List my tickets (subscriber)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: ticket_type
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketListResponse" }